feat: функционал ленты событий

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FeedStorage.java
@@ -1,14 +1,10 @@
 package ru.yandex.practicum.filmorate.storage;
 
 import ru.yandex.practicum.filmorate.model.Feed;
-import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.model.Genre;
 
-import java.util.Collection;
 import java.util.List;
-import java.util.Set;
 
 public interface FeedStorage {
 
-    List<Feed> getFeed(long userId);
+    List<Feed> getFeed(int userId);
 }
